// src/app/shared/interfaces/IADDRESS.ts
export interface IADDRESS {
  apartment: string
  block: string
  city: string
  detailsAboutAddress: string
  entry: string
  firstName: string
  floor: string
  id: number
  lastName: string
  municipality: string
  neighborhood: string
  phoneNumber: string
  street: string
  streetNumber: string
  zip: string

}